5730 Move DataFusionCompactionResult to its own file
